+++ server/graphql/invoice_line/src/mutations/outbound_shipment_line/unallocated_line/insert.rs
@@ -19,7 +19,7 @@ pub struct InsertInput {
     pub id: String,
     pub invoice_id: String,
     pub item_id: String,
-    pub quantity: u32,
+    pub quantity: f64,
 }
 
 #[derive(Interface)]
